requirements.txt
aiohttp>=3.9.0
chromadb>=0.4.0
duckduckgo-search>=4.1.1
numpy>=1.24.0
requests>=2.31.0
uvicorn>=0.24.0
fastapi>=0.104.0
python-dotenv>=1.0.0
beautifulsoup4==4.12.2
httpx>=0.27,<0.29
tabulate==0.9.0
streamlit==1.29.0
flask
flask-cors
pytest
pytest-asyncio
ollama
flask-restx